updated pictures, still need to link to pages, add routes and review design with team
