--- client/src/Components/Card/index.js
@@ -1,22 +1,0 @@
-// import React, { Component } from "react";
-// import Card from 'react-bootstrap/Card';
-// import cardImg from '../../Images/card.png';
-
-// export default class index extends Component {
-//   render() {
-//     return (
-//       <div>
-//         <Card className="bg-dark text-white 100px270">
-//           <Card.Img src={cardImg} alt="Card image" />
-//           <Card.ImgOverlay>
-//             <Card.Title>{this.props}</Card.Title>
-//             <Card.Text>
-              
-//             </Card.Text>
-//             <Card.Text>Last updated 3 mins ago</Card.Text>
-//           </Card.ImgOverlay>
-//         </Card>
-//       </div>
-//     );
-//   }
-// }
